requirements.txt
pytube
pydub 
ffmpeg
sshtunnel 
mysql-connector-python